Fix most tests
