+++ Project.toml
@@ -59,4 +59,4 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "ForwardDiff", "Manifolds", "Plots", "LineSearches"]
+test = ["Test", "ForwardDiff", "Manifolds", "Plots", "LineSearches", "LRUCache"]
